\name{print.CopyNumber450kCancer_data}
\alias{print.CopyNumber450kCancer_data}
\title{
Print Function for CopyNumber450kCancer Data(x)
}
\description{
Print function.
}
\arguments{
  \item{x}{
    Data object, as returned by \code{\link{ReadData}}.
  }
  \item{\dots}{
    Ignored.
  }
}

