\name{print.CopyNumber450kCancer_data}
\alias{print.CopyNumber450kCancer_data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Function for CopyNumber450kCancer Data(x)
}
\description{
Print function.
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    Data object, as returned by \code{\link{ReadData}}.
  }
  \item{\dots}{
    Ignored.
  }
}

